[ADD] Reestructuring folders, profile and character views and edit character functionality

+++ routes/web.php
@@ -13,7 +13,7 @@ use Illuminate\Support\Facades\Route;
 |
 */
 Route::get('/', 'App\Http\Controllers\HomeController@index')->name('home.index');
-Route::get('/items', 'App\Http\Controllers\UserItemController@index')->name('user.item.index');
+Route::get('/items', 'App\Http\Controllers\User\UserItemController@index')->name('user.item.index');
 /*
 |--------------------------------------------------------------------------
 | Here you will found all the routes related to admin role. This middleware
@@ -21,39 +21,39 @@ Route::get('/items', 'App\Http\Controllers\UserItemController@index')->name('use
 | is going to be rejected due an unauthorized request.
 */
 Route::middleware(['auth', 'admin'])->group(function () {
-    Route::get('/admin', 'App\Http\Controllers\AdminController@index')->name('admin.index');
+    Route::get('/admin', 'App\Http\Controllers\Admin\AdminController@index')->name('admin.index');
 
-    Route::get('/admin/items/create', 'App\Http\Controllers\ItemController@create')->name('admin.item.create');
-    Route::post('/admin/items/save', 'App\Http\Controllers\ItemController@save')->name('admin.item.save');
-    Route::delete('/admin/items/delete/{id}', 'App\Http\Controllers\ItemController@delete')->name('admin.item.delete');
-    Route::get('/admin/items', 'App\Http\Controllers\ItemController@index')->name('admin.item.index');
-    Route::get('/admin/items/{id}', 'App\Http\Controllers\ItemController@show')->name('admin.item.show');
-    Route::put('/admin/items/{id}', 'App\Http\Controllers\ItemController@update')->name('admin.item.update');
+    Route::get('/admin/items/create', 'App\Http\Controllers\Admin\AdminItemController@create')->name('admin.item.create');
+    Route::post('/admin/items/save', 'App\Http\Controllers\Admin\AdminItemController@save')->name('admin.item.save');
+    Route::delete('/admin/items/delete/{id}', 'App\Http\Controllers\Admin\AdminItemController@delete')->name('admin.item.delete');
+    Route::get('/admin/items', 'App\Http\Controllers\Admin\AdminItemController@index')->name('admin.item.index');
+    Route::get('/admin/items/{id}', 'App\Http\Controllers\Admin\AdminItemController@show')->name('admin.item.show');
+    Route::put('/admin/items/{id}', 'App\Http\Controllers\Admin\AdminItemController@update')->name('admin.item.update');
 
-    Route::get('/admin/reviews', 'App\Http\Controllers\ReviewController@index')->name('admin.review.index');
-    Route::get('/admin/reviews/{id}', 'App\Http\Controllers\ReviewController@show')->name('admin.review.show');
-    Route::delete('/admin/reviews/{id}', 'App\Http\Controllers\ReviewController@delete')->name('admin.review.delete');
+    Route::get('/admin/reviews', 'App\Http\Controllers\Admin\AdminReviewController@index')->name('admin.review.index');
+    Route::get('/admin/reviews/{id}', 'App\Http\Controllers\Admin\AdminReviewController@show')->name('admin.review.show');
+    Route::delete('/admin/reviews/{id}', 'App\Http\Controllers\Admin\AdminReviewController@delete')->name('admin.review.delete');
 
-    Route::get('/admin/users', 'App\Http\Controllers\AdminUserController@index')->name('admin.user.index');
-    Route::delete('/admin/users/{id}', 'App\Http\Controllers\AdminUserController@delete')->name('admin.user.delete');
+    Route::get('/admin/users', 'App\Http\Controllers\Admin\AdminUserController@index')->name('admin.user.index');
+    Route::delete('/admin/users/{id}', 'App\Http\Controllers\Admin\AdminUserController@delete')->name('admin.user.delete');
 });
 
 Route::middleware(['auth', 'user'])->group(function () {
     //user
-    Route::get('/myProfile', 'App\Http\Controllers\UserController@profile')->name('user.profile');
+    Route::get('/myProfile', 'App\Http\Controllers\User\UserController@profile')->name('user.profile');
 
-    Route::get('/my-purchases', 'App\Http\Controllers\UserMyPurchasesController@index')->name('user.purchase.index');
+    Route::get('/my-purchases', 'App\Http\Controllers\User\UserMyPurchasesController@index')->name('user.purchase.index');
 
-    Route::get('/character', 'App\Http\Controllers\CharacterController@editView')->name('character.editView');
-    Route::put('/character/update', 'App\Http\Controllers\CharacterController@update')->name('character.update');
+    Route::get('/character', 'App\Http\Controllers\User\UserCharacterController@editView')->name('character.editView');
+    Route::put('/character/update', 'App\Http\Controllers\User\UserCharacterController@update')->name('character.update');
 
-    Route::get('/items/{id}', 'App\Http\Controllers\UserItemController@show')->name('user.item.show');
-    Route::post('/items/{itemId}/review', 'App\Http\Controllers\UserItemController@addReview')->name('user.item.review');
+    Route::get('/items/{id}', 'App\Http\Controllers\User\UserItemController@show')->name('user.item.show');
+    Route::post('/items/{itemId}/review', 'App\Http\Controllers\User\UserItemController@addReview')->name('user.item.review');
 
-    Route::get('/cart', 'App\Http\Controllers\UserCartController@index')->name('user.cart.index');
-    Route::get('/cart/delete', 'App\Http\Controllers\UserCartController@delete')->name('user.cart.delete');
-    Route::post('/cart/add/{id}', 'App\Http\Controllers\UserCartController@add')->name('user.cart.add');
-    Route::get('/cart/purchase', 'App\Http\Controllers\UserCartController@purchase')->name('user.cart.purchase');
+    Route::get('/cart', 'App\Http\Controllers\User\UserCartController@index')->name('user.cart.index');
+    Route::get('/cart/delete', 'App\Http\Controllers\User\UserCartController@delete')->name('user.cart.delete');
+    Route::post('/cart/add/{id}', 'App\Http\Controllers\User\UserCartController@add')->name('user.cart.add');
+    Route::get('/cart/purchase', 'App\Http\Controllers\User\UserCartController@purchase')->name('user.cart.purchase');
 });
 
 Auth::routes();
